Added check for empty string when getting the latest tag from github Added jq to required packages for the script Updated auto_upgrade command to check for instructions to erase the node DB or not

+++ script.sh
@@ -121,6 +121,8 @@ case "$1" in
   git_clone
   build_node
   
+  DBQUERY=$(curl --silent "https://api.github.com/repos/ElrondNetwork/elrond-go/releases/latest" | jq -r .body | grep '*This release should start with a new DB*' -c)
+  
   INSTALLEDNODES=$(cat /opt/node/.numberofnodes)
   
   #Run the update process for each node
@@ -129,10 +131,24 @@ case "$1" in
         UPDATEINDEX=$(( $i - 1 ))
         UPDATEWORKDIR="/opt/node/node-$UPDATEINDEX"
         sudo cp $UPDATEWORKDIR/config/prefs.toml $UPDATEWORKDIR/config/prefs.toml.save
-        sudo systemctl stop elrond-node-$UPDATEINDEX
-        update
-        sudo mv $UPDATEWORKDIR/config/prefs.toml.save $UPDATEWORKDIR/config/prefs.toml && sudo chown node:node $UPDATEWORKDIR/config/prefs.toml
-        sudo systemctl start elrond-node-$UPDATEINDEX
+        
+        if [ "$DBQUERY" -eq "1" ]; then
+                        sudo systemctl stop elrond-node-$UPDATEINDEX
+                        echo "Database Cleanup Called ! Erasing... " >> $HOME/autoupdate.status        
+                        cleanup
+                        update
+                        sudo mv $UPDATEWORKDIR/config/prefs.toml.save $UPDATEWORKDIR/config/prefs.toml && sudo chown node:node $UPDATEWORKDIR/config/prefs.toml
+                        sudo systemctl start elrond-node-$UPDATEINDEX
+        
+        
+                      else
+                        sudo systemctl stop elrond-node-$UPDATEINDEX
+                        echo "Database Cleanup Not Needed ! Moving to next step... " >> $HOME/autoupdate.status
+                        update
+                        sudo mv $UPDATEWORKDIR/config/prefs.toml.save $UPDATEWORKDIR/config/prefs.toml && sudo chown node:node $UPDATEWORKDIR/config/prefs.toml
+                        sudo systemctl start elrond-node-$UPDATEINDEX
+        
+                      fi
       done
   ;;
